requirements.txt
streamlit
setuptools
# utilities 
numpy
pandas 

# Face recognition
scikit-learn
dlib-bin

git+https://github.com/ageitgey/face_recognition_models


supabase
bcrypt

segno
pillow



librosa
resemblyszer